moved function to better place
