+++ emiproc/exports/utils.py
@@ -3,42 +3,10 @@ import pandas as pd
 import xarray as xr
 
 from emiproc.inventories import Inventory
-from emiproc.profiles import temporal_profiles
+from emiproc.profiles.temporal_profiles import  get_index_in_profile
 from emiproc.utils.translators import inv_to_xarray
 
 
-def get_index_in_profile(
-    profile: temporal_profiles.TemporalProfile, time_range: pd.DatetimeIndex
-) -> pd.Series:
-    """Get the index in the profile for each time in the time range.
-
-    :param profile: the profile to use
-    :param time_range: the time range to use
-    :return: the index in the profile for each time in the time range
-    """
-
-    if profile == temporal_profiles.MounthsProfile:
-        indexes = time_range.month - 1
-    elif profile == temporal_profiles.DayOfYearProfile:
-        indexes = time_range.day_of_year - 1
-    elif profile == temporal_profiles.DailyProfile:
-        indexes = time_range.hour
-    elif profile == temporal_profiles.WeeklyProfile:
-        indexes = time_range.day_of_week
-    elif profile in [
-        temporal_profiles.HourOfYearProfile,
-        temporal_profiles.HourOfLeapYearProfile,
-    ]:
-        indexes = time_range.hour + (time_range.day_of_year - 1) * 24
-    elif profile == temporal_profiles.Hour3OfDayPerMonth:
-        indexes = (time_range.hour // 3) + (time_range.month - 1) * 8
-    else:
-        raise ValueError(f"Profile type {profile} not recognized")
-
-    assert indexes.min() >= 0
-    assert indexes.max() < profile.size
-
-    return indexes
 
 
 def get_temporally_scaled_array(
